Input flow

+++ src/quint/cabs/beans/Cab.java
@@ -5,9 +5,9 @@ public class Cab {
 	/** */
 	private int cabId;
 	
-	private int x;
+	private double x;
 	
-	private int y;
+	private double y;
 	
 	private String status;
 	
@@ -32,28 +32,28 @@ public class Cab {
 	/**
 	 * @return the x
 	 */
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
 	/**
 	 * @param x the x to set
 	 */
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
 	/**
 	 * @return the y
 	 */
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
 	/**
 	 * @param y the y to set
 	 */
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
@@ -98,8 +98,8 @@ public class Cab {
 	public void setCabType(String cabType) {
 		this.cabType = cabType;
 	}
-	
-	
+
+
 	
 	
 }
